# contracts/matterhorn.openapi.yaml
openapi: 3.0.3
info:
  title: Matterhorn REST API
  description: >-
    Thin REST facade over the Matterhorn actor model — a request/response view of the
    same state exposed on MQTT (nothing lives in one that isn't in the other). The HTTP
    verb carries intent: GET reads, PATCH applies a partial state change.
  version: 1.0.0

servers:
  - url: /

security:
  - ApiKey: []

paths:
  /api/devices:
    get:
      operationId: listDevices
      summary: List all known devices.
      tags: [devices]
      responses:
        '200':
          description: Every device and its exposes contract.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'

  /api/devices/{name}:
    get:
      operationId: getDeviceState
      summary: Read a device's current state.
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      responses:
        '200':
          description: The device's current retained state (same shape as the MQTT state topic).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceState'
        '404':
          description: No device with that friendly name.
    patch:
      operationId: setDeviceState
      summary: Apply a partial state change to a device.
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRequest'
      responses:
        '202':
          description: Accepted; the change is applied asynchronously against the fabric.
    delete:
      operationId: removeDevice
      summary: Un-pair (decommission) a device from the fabric.
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      responses:
        '202':
          description: Decommission started; the device leaves the fabric asynchronously.
        '404':
          description: No device with that friendly name.

  /api/devices/{name}/rename:
    post:
      operationId: renameDevice
      summary: Rename a device.
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameRequest'
      responses:
        '200':
          description: Renamed.
        '400':
          description: The requested name is not usable.
        '404':
          description: No device with that friendly name.
        '409':
          description: A device with the target name already exists.

  /api/commission:
    post:
      operationId: commission
      summary: Commission a new device by setup code.
      tags: [bridge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionRequest'
      responses:
        '202':
          description: Commissioning started; correlate progress by transaction id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionAccepted'

  /api/groups:
    get:
      operationId: listGroups
      summary: List all groups.
      tags: [groups]
      responses:
        '200':
          description: Every group and its member friendly names.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'

  /api/groups/{name}:
    put:
      operationId: putGroup
      summary: Create or replace a group.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMembers'
      responses:
        '200':
          description: Replaced.
        '201':
          description: Created.
        '400':
          description: The requested name is not usable.
        '409':
          description: The name collides with an existing device.
    patch:
      operationId: setGroupState
      summary: Fan out a partial state change to a group's members.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRequest'
      responses:
        '202':
          description: Accepted; applied asynchronously.
    delete:
      operationId: deleteGroup
      summary: Delete a group.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      responses:
        '202':
          description: Deleted.
        '404':
          description: No group with that name.

  /api/groups/{name}/members/{device}:
    put:
      operationId: addGroupMember
      summary: Add a device to a group.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
        - name: device
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Added.
        '404':
          description: No such group or device.
    delete:
      operationId: removeGroupMember
      summary: Remove a device from a group.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
        - name: device
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed.
        '404':
          description: No group with that name.

  /api/groups/{name}/rename:
    post:
      operationId: renameGroup
      summary: Rename a group.
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameRequest'
      responses:
        '200':
          description: Renamed.
        '400':
          description: Not usable.
        '404':
          description: No group with that name.
        '409':
          description: Target name taken.

  /api/scenes:
    get:
      operationId: listScenes
      summary: List all scenes.
      tags: [scenes]
      responses:
        '200':
          description: Every scene and its member friendly names.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scene'

  /api/scenes/{name}:
    put:
      operationId: putScene
      summary: Store (create-or-replace) a scene — snapshot members or use explicit state.
      tags: [scenes]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreSceneRequest'
      responses:
        '200':
          description: Replaced.
        '201':
          description: Created.
        '400':
          description: Not usable.
    delete:
      operationId: deleteScene
      summary: Delete a scene.
      tags: [scenes]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      responses:
        '202':
          description: Deleted.
        '404':
          description: No scene with that name.

  /api/scenes/{name}/recall:
    post:
      operationId: recallScene
      summary: Recall a scene (fan out its stored values).
      tags: [scenes]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      responses:
        '202':
          description: Accepted; applied asynchronously.
        '404':
          description: No scene with that name.

  /api/scenes/{name}/rename:
    post:
      operationId: renameScene
      summary: Rename a scene.
      tags: [scenes]
      parameters:
        - $ref: '#/components/parameters/FriendlyName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameRequest'
      responses:
        '200':
          description: Renamed.
        '400':
          description: Not usable.
        '404':
          description: No scene with that name.
        '409':
          description: Target name taken.

  /api/bridge/info:
    get:
      operationId: getBridgeInfo
      summary: Service metadata.
      tags: [bridge]
      responses:
        '200':
          description: Bridge/service information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeInfo'

components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key

  parameters:
    FriendlyName:
      name: name
      in: path
      required: true
      description: The device friendly name.
      schema:
        type: string

  schemas:
    Device:
      type: object
      required: [friendly_name, node_id, endpoint, device_type, reachable, exposes]
      properties:
        friendly_name: { type: string }
        node_id: { type: string }
        endpoint: { type: integer, format: int32 }
        vendor_name: { type: string, nullable: true }
        product_name: { type: string, nullable: true }
        vendor_id: { type: integer, format: int32 }
        product_id: { type: integer, format: int32 }
        device_type: { type: string }
        reachable: { type: boolean }
        transport: { type: string, description: 'wifi | thread | ethernet | unknown' }
        exposes:
          type: array
          items:
            $ref: '#/components/schemas/Expose'

    Expose:
      type: object
      required: [type, property, access]
      properties:
        type: { type: string, description: 'binary | numeric | enum' }
        property: { type: string }
        access: { type: integer, format: int32, description: 'Z2M bitmask: 1=published, 2=set, 4=get' }
        value_on: { type: string, nullable: true }
        value_off: { type: string, nullable: true }
        value_min: { type: integer, format: int32, nullable: true }
        value_max: { type: integer, format: int32, nullable: true }
        unit: { type: string, nullable: true }
        values:
          type: array
          nullable: true
          items: { type: string }

    SetRequest:
      type: object
      additionalProperties: true
      description: >-
        One or more settable properties; omitted properties are left unchanged. Free-form: the same
        open vocabulary as the MQTT `<base>/<name>/set` payload (e.g. state, brightness, color_temp,
        hue, saturation, color_x, color_y, position, fan_mode, percent, system_mode,
        occupied_heating_setpoint, occupied_cooling_setpoint, transition, identify) — the accepted
        keys are whatever the device's `exposes` advertise as settable.

    DeviceState:
      type: object
      description: Free-form map of semantic property names to current values.
      additionalProperties: true

    CommissionRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Matter setup code (QR 'MT:' string or 11-digit manual code).

    RenameRequest:
      type: object
      required: [to]
      properties:
        to:
          type: string
          description: The new friendly name (slugified server-side into a legal topic segment).

    CommissionAccepted:
      type: object
      required: [transaction]
      properties:
        transaction: { type: string }

    BridgeInfo:
      type: object
      required: [service]
      properties:
        service: { type: string }
        thread:
          $ref: '#/components/schemas/ThreadStatus'

    ThreadStatus:
      type: object
      description: >-
        Whether a brand-new Thread device can be commissioned right now. Needs both Thread
        credentials to hand the device and Bluetooth on the controller to hand them over with.
        Irrelevant for Wi-Fi/Ethernet devices, which are commissioned over the network.
      required: [available, source]
      properties:
        available: { type: boolean }
        source:
          type: string
          description: 'Where the Thread credentials came from: otbr | configured | none.'
        border_router:
          type: string
          description: 'Border router the credentials were read from.'
        reason:
          type: string
          description: 'Why Thread commissioning is unavailable, in plain language. Absent when it is available.'

    Group:
      type: object
      required: [friendly_name, members]
      properties:
        friendly_name: { type: string }
        members: { type: array, items: { type: string } }

    GroupMembers:
      type: object
      properties:
        members: { type: array, items: { type: string }, description: 'Initial member friendly names.' }

    Scene:
      type: object
      required: [friendly_name, members]
      properties:
        friendly_name: { type: string }
        members: { type: array, items: { type: string } }

    StoreSceneRequest:
      type: object
      properties:
        devices: { type: array, items: { type: string }, description: 'Members to snapshot (ignored when state is present).' }
        state:
          type: object
          additionalProperties: { type: object, additionalProperties: true }
          description: 'Explicit per-device property map; when present, stored verbatim instead of snapshotting.'